Refactor: Chatbot 관련 컴포넌트 구조 변경
- ChatItem을 Compound Component 구조로 변경
