+++ src/components/organisms/HeadlineBanner/index.ts
@@ -1,0 +1,2 @@
+export * from './HeadlineBanner';
+export * from './HeadlineBannerSkeleton';
