+++ src/types/common/headline.ts
@@ -1,5 +1,7 @@
+import { PostTypes } from '@/types/post/postType';
+
 export interface Headline {
   id: number;
   title: string;
-  type: 'notice' | 'news';
+  type: PostTypes;
 }
